+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed table cell styling in `markdown` view
 - Added thousandth delimiter for default rendering of a number in `table` view cell
 - Fixed `popup` view to auto hide when trigger elements are removing from a document
+- Tweaked styles for nested list views
 
 ## 1.0.0-beta.62 (14-10-2021)