+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## next
 
-- Fixed table cell styling in `markdown` view
-- Added thousandth delimiter for default rendering of a number in `table` view cell
 - Fixed `popup` view to auto hide when trigger elements are removing from a document
+- Fixed table cell styling in `markdown` view
+- Tweaked styles for list and checkbox in `markdown` view
 - Tweaked styles for nested list views
+- Added thousandth delimiter for default rendering of a number in `table` view cell
 
 ## 1.0.0-beta.62 (14-10-2021)