+++ CHANGELOG.md
@@ -1,11 +1,12 @@
-## next
+## 1.0.0-beta.63 (07-02-2022)
 
+- Bumped dependencies to latest versions (fixes security issues)
 - Fixed `popup` view to auto hide when trigger elements are removing from a document
 - Fixed progress bar state update on data loading after a discovery instance init
 - Removed `App.modelfreeLoadData()` method, use `App#loadDataFromEvent()` or `#.actions.upload` instead
 - Added `Widget#unloadData()` method to reset loaded data and context if any
 - Added `#.dataLoaded` property to indicate is data loaded or not
-- Generalized approach to load data:
+- Generalized approach to upload data:
     - API exposed as an extension which can be imported as `import { upload } from '@discoveryjs/discovery'` and configured via `options.upload`
     - Turns on automatically when `options.upload` is truthy for `App`
     - Options supported for `options.upload`: